% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_get.R
\name{bh_get}
\alias{bh_get}
\title{Get data from banquehydro}
\usage{
bh_get(procedure = "QTVAR", station, t1, t2, sleep = 10)
}
\arguments{
\item{procedure}{procedure à importer : au choix pour l'instant "QTVAR","QJM" (QTVAR par défaut)}

\item{station}{Code de la station}

\item{t1}{Date de début d'événement au format string "JJ/MM/AAAA HH:MM:SS" si procedure=QTVAR ou au format "AAAA" si procedure=QJM}

\item{t2}{Date de fin d'événement au format string "JJ/MM/AAAA HH:MM:SS"si procedure=QTVAR ou au format "AAAA" si procedure=QJM}
}
\value{
dataframe contenant le tableau produit sur la page de visualisation de la chronique de la banque hydro
}
\description{
Collect data from banquehydro stating procedure type, station code and beginning/ending times
}
\examples{
df_qtvar<-bh_get(procedure="QTVAR",
                 station="Y3204010",
                 t1= "16/03/2017 00:00",
                 t2= "17/03/2017 23:59")


df_qjm <- bh_get(procedure="QJM",
           station="Y3204010",
           t1= "2005",
           t2= "2006",
           sleep=2)
}
