% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_get_qjm.R
\name{bh_get_qjm}
\alias{bh_get_qjm}
\title{get QJM data for a station between two times t1 and t2}
\usage{
bh_get_qjm(station, t1, t2)
}
\arguments{
\item{station}{station code}

\item{t1}{Beginning year}

\item{t2}{Ending year}
}
\description{
This is used to get QJM data between two years t1 and t2.
}
\examples{
df_ts<-bh_get_qtvar(station="V2942010",
                    t1=2008,
                    t2=2010)
}
