% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_get_qtvar.R
\name{bh_get_qtvar}
\alias{bh_get_qtvar}
\title{Used from bh_get to get QTVAR data}
\usage{
bh_get_qtvar(t1, t2)
}
\arguments{
\item{t1}{Date de début d'événement au format "%j/%M/%Y %h:%m"}

\item{t2}{Date de fin d'événement au format "%j/%M/%Y %h:%m"}
}
\value{
dataframe contenant le tableau produit sur la page de visualisation de la chronique de la banque hydro
}
\description{
This is used to get QTVAR data between two times.
}
\examples{
df_qtvar<-bh_get_qtvar("16/03/2017 00:00", "05/04/2017 23:59")
}
