% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_qjm.R
\name{collect_qjm}
\alias{collect_qjm}
\title{collect QTVAR data from response page, for a given year}
\usage{
collect_qjm(res, year)
}
\arguments{
\item{res}{response page}

\item{year}{year}
}
\description{
collect QTVAR data from response page, for a given year
}
