% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_qtvar.R
\name{collect_qtvar}
\alias{collect_qtvar}
\title{collect QTVAR data from response page}
\usage{
collect_qtvar(res)
}
\arguments{
\item{res}{response page}
}
\description{
collect QTVAR data from response page
}
