% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_time_for_qtvar.R
\name{format_time_for_qtvar}
\alias{format_time_for_qtvar}
\title{Formats time as "\%d/\%m/\%Y \%H:\%M"}
\usage{
format_time_for_qtvar(mytime)
}
\arguments{
\item{mytime}{time to be formatted}
}
\value{
formatted time
}
\description{
Formats time as "\%d/\%m/\%Y \%H:\%M"
}
