% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_to_qjm.R
\name{get_to_QJM}
\alias{get_to_QJM}
\title{#' @description This is used to get QTVAR data between two times.}
\usage{
get_to_QJM(year)
}
\description{
#' @description This is used to get QTVAR data between two times.
}
