% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_to_qtvar.R
\name{get_to_qtvar}
\alias{get_to_qtvar}
\title{get_to_qtvar from webpage}
\usage{
get_to_qtvar(res, t1, t2)
}
\arguments{
\item{res}{page}

\item{t1}{starting time formatted as "\%j/\%M/\%Y \%h:\%m"}

\item{t2}{ending time formatted as "\%j/\%M/\%Y \%h:\%m"}
}
\value{
page with dataframe of qtvar for station between t1 and t2 as visualized on banquehydro
}
\description{
This function is used by bh_get_qtvar() to get page of results for QTVAR
}
