% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_to_station.R
\name{get_to_station}
\alias{get_to_station}
\title{get_to_station from webpage}
\usage{
get_to_station(station)
}
\arguments{
\item{station}{gauging station code}
}
\value{
page corresponding to station on banquehydro
}
\description{
get_to_station from webpage
}
