% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_get_qjm.R
\name{bh_get_qjm}
\alias{bh_get_qjm}
\title{Used from bh_get to get QJM data}
\usage{
bh_get_qjm(year, sleep)
}
\arguments{
\item{year}{the year for which the data is collected}

\item{sleep}{how much time the system waits between two GET requests (in seconds)}
}
\value{
QJM for year `year`
}
\description{
This is used from bh_get to get QJM data between two times.
}
\examples{
df_ts<-bh_get_qjm(year="2018", sleep=1)
}
