% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bh_get_qtvar.R
\name{bh_get_qtvar}
\alias{bh_get_qtvar}
\title{Get QTVAR data for a station between two times t1 and t2}
\usage{
bh_get_qtvar(station, t1, t2)
}
\arguments{
\item{station}{station code}

\item{t1}{starting time formatted as "\%j/\%M/\%Y \%h:\%m"}

\item{t2}{ending time formatted as "\%j/\%M/\%Y \%h:\%m"}
}
\value{
tibble with QTVAR data
}
\description{
This is used to get QTVAR data between two times.
}
\examples{
df_qtvar<-bh_get_qtvar(station="V2942010",
                       t1="05/02/2015 15:00",
                       t2="06/02/2015 18:00")
}
