% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_qjm.R
\name{collect_qjm}
\alias{collect_qjm}
\title{collect_qjm}
\usage{
collect_qjm(res, year)
}
\arguments{
\item{res}{response page}

\item{year}{year}
}
\value{
QJM data corresponding to year
}
\description{
collect_qjm
}
