% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_qtvar.R
\name{collect_qtvar}
\alias{collect_qtvar}
\title{collect_qtvar from webpage}
\usage{
collect_qtvar(res)
}
\arguments{
\item{res}{response page}
}
\value{
QTVAR data corresponding to page
}
\description{
collect_qtvar from webpage
}
