% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_time_for_qtvar.R
\name{format_time_for_qtvar}
\alias{format_time_for_qtvar}
\title{Format time to be formatted as banqueHydro expects for QTVAR ("\%d/\%m/\%Y \%H:\%M")}
\usage{
format_time_for_qtvar(mytime)
}
\arguments{
\item{mytime}{time}
}
\value{
time in format "%d/%m/%Y %H:%M"
}
\description{
Format time to be formatted as banqueHydro expects for QTVAR ("\%d/\%m/\%Y \%H:\%M")
}
