% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_to_procedure.R
\name{get_to_procedure}
\alias{get_to_procedure}
\title{get_to_procedure}
\usage{
get_to_procedure(res, procedure, station)
}
\arguments{
\item{res}{page}

\item{procedure}{procedure: import either "QTVAR" or "QJM"}

\item{station}{gauging station code}
}
\value{
page corresponding to procedure and station
}
\description{
get_to_procedure
}
