% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_to_qjm.R
\name{get_to_qjm}
\alias{get_to_qjm}
\title{get_to_qjm}
\usage{
get_to_qjm(res, year)
}
\arguments{
\item{res}{page}

\item{year}{year for which data should be collected}
}
\value{
page with dataframe of QJM for specified year as visualised on banquehydro
}
\description{
This function is used by bh_get_qjm() to get page of results for QJM
}
