% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_to_qtvar.R
\name{get_to_qtvar}
\alias{get_to_qtvar}
\title{Used from bh_get to get QTVAR data}
\usage{
get_to_qtvar(station, t1, t2)
}
\arguments{
\item{station}{gauging station}

\item{t1}{Date de début d'événement au format "%j/%M/%Y %h:%m"}

\item{t2}{Date de fin d'événement au format "%j/%M/%Y %h:%m"}
}
\value{
dataframe contenant le tableau produit sur la page de visualisation de la chronique de la banque hydro
}
\description{
This is used to get QTVAR data between two times.
}
\examples{
df_qtvar<-bh_get_qtvar("16/03/2017 00:00", "18/03/2017 23:59")
}
